+++ src/main/cpp/RobotContainer.cpp
@@ -31,7 +31,6 @@
 #include "commands/IntakeOutCommand.h"
 #include "subsystems/ClimbSubsystem.h"
 #include "subsystems/DriveSubsystem.h"
-#include "utils/CommandUtils.h"
 #include "utils/InputUtils.h"
 #include "utils/ShuffleboardLogger.h"
