+++ src/main/include/subsystems/BaseSingleAxisSubsystem.h
@@ -158,8 +158,7 @@ class BaseSingleAxisSubsystem : public ISingleAxisSubsystem {
    * @param speed Percentage speed
    */
   void RunMotorExternal(double speed, bool ignoreEncoder = false) override {
-    // TODO: constant
-    if (abs(speed) <= 0.05) {
+    if (abs(speed) <= kMotorDeadSpeedRange) {
       if (_isMovingToPosition)
         return;  // Don't set the motor and overwrite a potential
                  // automated movement
@@ -240,7 +239,7 @@ class BaseSingleAxisSubsystem : public ISingleAxisSubsystem {
 
     // TODO: Constant wrap-around value
     if (GetCurrentPosition() <= _config.minDistance ||
-        GetCurrentPosition() >= 350.0) {
+        GetCurrentPosition() >= kMaxRotPosition) {
       if (_log)
         ConsoleLogger::getInstance().logInfo(_prefix, "AT HOME ENCODER %s", "");
       return true;
