+++ src/main/include/subsystems/ClimbSubsystem.h
@@ -22,8 +22,6 @@ class ClimbSubsystem
 
   double GetCurrentPosition() override {
     auto curPosition = _enc.GetPosition() * _config.distancePerRevolution;
-    // ConsoleLogger::getInstance().logVerbose("Climbing Subsystem", "current position
-    // %f", curPosition);
     return curPosition;
   }
