+++ src/main/include/subsystems/IntakeSubsystem.h
@@ -58,21 +58,4 @@ class IntakeSubsystem : public frc2::SubsystemBase {
       IntakingConstants::kUpperAmpBeamBreakDigitalPort};
   frc::DigitalInput m_lowerAmpBeamBreak{
       IntakingConstants::kLowerampBeamBreakDigitalPort};
-
-  //   rev::SparkPIDController m_rightIntakeSpinnyBoyPID =
-  //       m_rightIntakeSpinnyBoy.GetPIDController();
-  //   rev::SparkPIDController m_leftIntakeSpinnyBoyPID =
-  //       m_leftIntakeSpinnyBoy.GetPIDController();
-
-  //   PidSettings intakePidSettings = {.p = IntakingPID::kIntakingP,
-  //                                    .i = IntakingPID::kIntakingI,
-  //                                    .d = IntakingPID::kIntakingD,
-  //                                    .iZone = IntakingPID::kIntakingIZone,
-  //                                    .ff = IntakingPID::kIntakingFF};
-
-  //   PidMotorControllerPair intakeMotors{"Intake", m_leftIntakeSpinnyBoyPID,
-  //                                       m_rightIntakeSpinnyBoyPID,
-  //                                       intakePidSettings, kMaxRpm};
-
-  //   PidMotorControllerPairTuner intakeTuner{intakeMotors};
 };
